dashboard/requirements.txt
streamlit
psycopg2-binary==2.9.9
